Extract a base API resource
